+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 **I'm using for some apps the same icons. If this is the case with your apps or if something other went wrong, just report it [here](https://github.com/scoute-dich/Baumann_Theme/issues) or on the [xda threat](http://forum.xda-developers.com/android/themes/cm12-theme-source-t3164482).**
 
+
+### 4.5.2
+
+- fixed: f-droid build failure
+
+
+### 4.5.1
+
+- fixed: app selection icon request
+
+
+### 4.5
+
+- added: icons (thanks to @thomas31415 on xda thread)
+- added: new wallpapers
+- added: new app icon
+- added: new boot animation
+- removed: some wallpapers
+- improved: wallpaper fragment layout
+- improved: icon request fragment layout
+- fixed: fab behavior
+- fixed: some force closes on Marshmallow+
+
+
 ### 4.4
 
 - added: icons (thanks to @thomas31415 on xda thread)